Move geodata_helpers to utils for easier access from modules.
